+++ Spring boot/books/books/src/main/java/pl/edu/pwr/ztw/books/books/IBooksService.java
@@ -4,8 +4,8 @@ import java.util.Collection;
 
 public interface IBooksService {
     Collection<Book> getBooks();
-    Book getBook(int id);
+    Book getBook(long id);
     Book addBook(Book book);
-    Book updateBook(int id, Book book);
-    boolean deleteBook(int id);
+    Book updateBook(long id, Book book);
+    boolean deleteBook(long id);
 }
